requirements.txt
# Core AutoGen and OpenAI (Updated for AutoGen 0.7.1)
autogen-agentchat>=0.7.1
autogen-ext[openai,azure]>=0.7.1

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.5.0

# Multi-Agent Frameworks
crewai>=0.28.8

# LangChain Integration (optional)
langchain>=0.1.0
langchain-openai>=0.0.8

# HTTP and Async Support
aiohttp>=3.9.0
requests>=2.31.0

# Utilities
tenacity>=8.2.0
typing-extensions>=4.8.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Optional: Advanced Features
# jupyter>=1.0.0  # For notebook examples
# matplotlib>=3.7.0  # For visualization
# plotly>=5.17.0  # For interactive plots